+++ common/src/main/java/io/github/mortuusars/envelope/mixin/spawners/ServerLevelMixin.java
@@ -1,16 +1,14 @@
 package io.github.mortuusars.envelope.mixin.spawners;
 
-import io.github.mortuusars.envelope.Config;
 import io.github.mortuusars.envelope.world.entity.spawner.BackgroundCourierSpawner;
 import io.github.mortuusars.envelope.world.entity.spawner.FinishedBackgroundCourierSpawner;
+import io.github.mortuusars.envelope.world.entity.spawner.DroppedMailSpawner;
 import io.github.mortuusars.envelope.world.entity.spawner.Spawner;
-import io.github.mortuusars.envelope.world.mail.MailService;
 import net.minecraft.core.Holder;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.profiling.ProfilerFiller;
-import net.minecraft.world.level.GameRules;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.dimension.DimensionType;
 import net.minecraft.world.level.storage.WritableLevelData;
@@ -29,7 +27,8 @@ public abstract class ServerLevelMixin extends Level {
     @Unique
     private final List<Spawner> envelope$spawners = List.of(
           new BackgroundCourierSpawner(),
-          new FinishedBackgroundCourierSpawner()
+          new FinishedBackgroundCourierSpawner(),
+          new DroppedMailSpawner()
     );
 
     protected ServerLevelMixin(WritableLevelData levelData, ResourceKey<Level> dimension, RegistryAccess registryAccess,
@@ -41,13 +40,6 @@ public abstract class ServerLevelMixin extends Level {
 
     @Inject(method = "tick", at = @At("RETURN"))
     private void onTick(BooleanSupplier hasTimeLeft, CallbackInfo ci) {
-        if (!MailService.operatesIn(this)) {
-            return;
-        }
-
-        if (!Config.Server.DELIVERY_SPAWNING_RESPECTS_DOMOBSPAWNING_RULE.get()
-              || getGameRules().getBoolean(GameRules.RULE_DOMOBSPAWNING)) {
-            envelope$spawners.forEach(spawner -> spawner.tick((ServerLevel) (Object) this));
-        }
+        envelope$spawners.forEach(spawner -> spawner.tick((ServerLevel) (Object) this));
     }
 }