+++ common/src/main/java/io/github/mortuusars/envelope/world/mail/delivery/PhysicalCourier.java
@@ -1,7 +1,7 @@
 package io.github.mortuusars.envelope.world.mail.delivery;
 
 import io.github.mortuusars.envelope.world.mail.delivery.background.BackgroundCourier;
-import io.github.mortuusars.envelope.world.entity.SpawnableEntityData;
+import io.github.mortuusars.envelope.world.entity.spawning.SpawnableEntityData;
 import io.github.mortuusars.envelope.world.mail.MailService;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.server.level.ServerLevel;
@@ -10,7 +10,7 @@ import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.phys.Vec3;
 
-public interface TransitionableCourier extends Courier {
+public interface PhysicalCourier extends Courier {
     SpawnableEntityData toSpawnableData();
 
     void setDelivery(Delivery delivery);